Simplify model catalog refresh entrypoint
