+++ website/docs/reference/model-catalog.md
@@ -58,12 +58,15 @@ Field notes:
 | When | What happens |
 |---|---|
 | `/model` or `hermes model` | Fetches live provider catalogs when authenticated; uses this manifest/cache as fallback |
+| `hermes model --refresh` | Forces a non-interactive refresh of models.dev, the hosted manifest, Ollama Cloud, and authenticated picker catalogs |
+| Daily model-refresh cron | Calls the same refresh path through `scripts/refresh_model_catalogs.py` and emits `{"wakeAgent": false}` so no LLM turn is used |
 | Disk cache fresh (< TTL) | No network hit |
 | Network failure with cache | Silent fallback to cache, one log line |
 | Network failure, no cache | Silent fallback to in-repo snapshot |
 | Manifest fails schema validation | Treated as unreachable |
 
 Cache location: `~/.hermes/cache/model_catalog.json`.
+Refresh status location: `~/.hermes/cache/model_catalog_refresh_status.json`.
 
 ## Config
